pubmed_downloader/requirements.txt
biopython==1.83
tqdm==4.66.2
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2 